backend/src/services/multimodal_service.py
"""
MultiModalService for image analysis, audio transcription, and rich content processing
"""

from typing import Dict, List, Optional, Any, Tuple
import base64
import io
from pathlib import Path

try:
    # Image processing
    from PIL import Image
    import torch
    from transformers import AutoProcessor, AutoModelForCausalLM

    # Audio processing
    import whisper

    MULTIMODAL_AVAILABLE = True
except ImportError:
    MULTIMODAL_AVAILABLE = False

from .ai_service import get_ai_service


class MultiModalService:
    """Service for multi-modal content processing (images, audio, rich content)"""

    def __init__(self):
        self.ai_service = get_ai_service()

        # Initialize models (lazy loading)
        self._image_model = None
        self._image_processor = None
        self._audio_model = None

    async def _ensure_image_model(self):
        """Lazy load image analysis model"""
        if not MULTIMODAL_AVAILABLE or self._image_model is not None:
            return

        try:
            # Use a lightweight vision model (placeholder - would use actual model)
            # In production, this would load something like LLaVA or similar
            print("Image model would be loaded here")
        except Exception as e:
            print(f"Failed to load image model: {e}")

    async def _ensure_audio_model(self):
        """Lazy load audio transcription model"""
        if not MULTIMODAL_AVAILABLE or self._audio_model is not None:
            return

        try:
            # Load Whisper model for transcription
            self._audio_model = whisper.load_model("base")
        except Exception as e:
            print(f"Failed to load audio model: {e}")

    async def analyze_image(self, image_path: str, query: Optional[str] = None) -> Dict[str, Any]:
        """Analyze image using vision-language model"""
        await self._ensure_image_model()

        try:
            # Placeholder implementation
            # In production, this would:
            # 1. Load and preprocess image
            # 2. Run inference with vision model
            # 3. Generate description/analysis

            mock_description = f"Analysis of image at {Path(image_path).name}"
            mock_objects = ["object1", "object2", "object3"]

            # Use AI service for any additional query processing
            if query:
                enhanced_query = f"Based on this image analysis, {query}"
                ai_response = await self.ai_service.generate_response(enhanced_query)
                answer = ai_response["response"]
            else:
                answer = None

            return {
                "description": mock_description,
                "objects": mock_objects,
                "answer": answer,
                "confidence": 0.85,
                "processing_time": 2.5
            }

        except Exception as e:
            return {
                "error": f"Image analysis failed: {str(e)}",
                "description": "Unable to analyze image",
                "objects": [],
                "confidence": 0.0,
                "processing_time": 0.0
            }

    async def analyze_image_from_bytes(self, image_bytes: bytes, query: Optional[str] = None) -> Dict[str, Any]:
        """Analyze image from byte data"""
        try:
            # Save temporarily for processing
            temp_path = f"/tmp/temp_image_{hash(image_bytes)}.jpg"
            with open(temp_path, "wb") as f:
                f.write(image_bytes)

            result = await self.analyze_image(temp_path, query)

            # Clean up temp file
            Path(temp_path).unlink(missing_ok=True)

            return result

        except Exception as e:
            return {
                "error": f"Image analysis failed: {str(e)}",
                "description": "Unable to analyze image",
                "objects": [],
                "confidence": 0.0,
                "processing_time": 0.0
            }

    async def transcribe_audio(self, audio_path: str, language: Optional[str] = None) -> Dict[str, Any]:
        """Transcribe audio file to text"""
        await self._ensure_audio_model()

        try:
            if not self._audio_model:
                return {
                    "error": "Audio transcription model not available",
                    "transcription": "",
                    "segments": [],
                    "language": "unknown",
                    "processing_time": 0.0
                }

            # Run transcription
            result = self._audio_model.transcribe(
                audio_path,
                language=language if language != "auto" else None,
                verbose=True
            )

            # Format segments
            segments = []
            for segment in result["segments"]:
                segments.append({
                    "start": segment["start"],
                    "end": segment["end"],
                    "text": segment["text"]
                })

            return {
                "transcription": result["text"],
                "segments": segments,
                "language": result.get("language", "unknown"),
                "processing_time": result.get("processing_time", 0.0)
            }

        except Exception as e:
            return {
                "error": f"Audio transcription failed: {str(e)}",
                "transcription": "",
                "segments": [],
                "language": "unknown",
                "processing_time": 0.0
            }

    async def transcribe_audio_from_bytes(self, audio_bytes: bytes, language: Optional[str] = None) -> Dict[str, Any]:
        """Transcribe audio from byte data"""
        try:
            # Save temporarily for processing
            temp_path = f"/tmp/temp_audio_{hash(audio_bytes)}.wav"
            with open(temp_path, "wb") as f:
                f.write(audio_bytes)

            result = await self.transcribe_audio(temp_path, language)

            # Clean up temp file
            Path(temp_path).unlink(missing_ok=True)

            return result

        except Exception as e:
            return {
                "error": f"Audio transcription failed: {str(e)}",
                "transcription": "",
                "segments": [],
                "language": "unknown",
                "processing_time": 0.0
            }

    def render_rich_content(self, content: str, content_type: str,
                           metadata: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
        """Render rich content for chat display"""
        try:
            metadata = metadata or {}

            if content_type == "markdown":
                # Process markdown (placeholder)
                rendered = f"<div class='markdown-content'>{content}</div>"

            elif content_type == "table":
                # Process table data (placeholder)
                rendered = f"<table class='chat-table'>{content}</table>"

            elif content_type == "code":
                # Syntax highlight code
                language = metadata.get("language", "text")
                rendered = f"<pre class='language-{language}'><code>{content}</code></pre>"

            elif content_type == "image":
                # Render image
                rendered = f"<img src='{content}' alt='Generated image' class='chat-image' />"

            else:
                # Plain text fallback
                rendered = f"<div class='chat-content'>{content}</div>"

            return {
                "renderedContent": rendered,
                "contentType": content_type,
                "metadata": metadata,
                "contentId": f"content_{hash(content)}"
            }

        except Exception as e:
            return {
                "error": f"Content rendering failed: {str(e)}",
                "renderedContent": f"<div class='error'>Failed to render content</div>",
                "contentType": content_type,
                "metadata": metadata or {}
            }

    async def extract_text_from_image(self, image_path: str) -> str:
        """Extract text from image using OCR (placeholder)"""
        try:
            # Placeholder for OCR implementation
            # In production, would use Tesseract or similar
            return f"Extracted text from {Path(image_path).name}"

        except Exception as e:
            return f"OCR failed: {str(e)}"

    def get_supported_formats(self) -> Dict[str, List[str]]:
        """Get supported formats for each modality"""
        return {
            "images": ["image/jpeg", "image/png", "image/gif", "image/webp"],
            "audio": ["audio/mpeg", "audio/wav", "audio/ogg", "audio/mp4"],
            "text": ["text/plain", "text/markdown", "application/pdf"]
        }


# Global instance for dependency injection
_multimodal_service_instance = None

def get_multimodal_service() -> MultiModalService:
    """Get singleton MultiModalService instance"""
    global _multimodal_service_instance
    if _multimodal_service_instance is None:
        _multimodal_service_instance = MultiModalService()
    return _multimodal_service_instance